NODE-312: Refactored HashSetCasperTest to allow generic test node creation.
